Async refactor in place (#7736)
* feat: allow both callback&and await

* feat: ignore async key

* feat: callbackify and promisify in same file

* Revert "feat: callbackify and promisify in same file"

This reverts commit cea206a9b8e6d8295310074b18cc82a504487862.

* feat: no need to store .callbackify

* feat: change getTopics to async

* feat: remove .async

* fix: byScore

* feat: rewrite topics/index and social with async/await

* fix: rewrite topics/data.js

fix issue with async.waterfall, only pass result if its not undefined

* feat: add callbackify to redis/psql

* feat: psql use await

* fix: redis :volcano:

* feat: less returns

* feat: more await rewrite

* fix: redis tests

* feat: convert sortedSetAdd

rewrite psql transaction to async/await

* feat: :dog:

* feat: test

* feat: log client and query

* feat: log bind

* feat: more logs

* feat: more logs

* feat: check perform

* feat: dont callbackify transaction

* feat: remove logs

* fix: main functions

* feat: more logs

* fix: increment

* fix: rename

* feat: remove cls

* fix: remove console.log

* feat: add deprecation message to .async usage

* feat: update more dbal methods

* fix: redis :voodoo:

* feat:  fix redis zrem, convert setObject

* feat: upgrade getObject methods

* fix: psql getObjectField

* fix: redis tests

* feat: getObjectKeys

* feat: getObjectValues

* feat: isObjectField

* fix: add missing return

* feat: delObjectField

* feat: incrObjectField

* fix: add missing await

* feat: remove exposed helpers

* feat: list methods

* feat: flush/empty

* feat: delete

* fix: redis delete all

* feat: get/set

* feat: incr/rename

* feat: type

* feat: expire

* feat: setAdd

* feat: setRemove

* feat: isSetMember

* feat: getSetMembers

* feat: setCount, setRemoveRandom

* feat: zcard,zcount

* feat: sortedSetRank

* feat: isSortedSetMember

* feat: zincrby

* feat: sortedSetLex

* feat: processSortedSet

* fix: add mising await

* feat: debug psql

* fix: psql test

* fix: test

* fix: another test

* fix: test fix

* fix: psql tests

* feat: remove logs

* feat: user arrow func

use builtin async promises

* feat: topic bookmarks

* feat: topic.delete

* feat: topic.restore

* feat: topics.purge

* feat: merge

* feat: suggested

* feat: topics/user.js

* feat: topics modules

* feat: topics/follow

* fix: deprecation msg

* feat: fork

* feat: topics/posts

* feat: sorted/recent

* feat: topic/teaser

* feat: topics/tools

* feat: topics/unread

* feat: add back node versions

disable deprecation notice
wrap async controllers in try/catch

* feat: use db directly

* feat: promisify in place

* fix: redis/psql

* feat: deprecation message

logs for psql

* feat: more logs

* feat: more logs

* feat: logs again

* feat: more logs

* fix: call release

* feat: restore travis, remove logs

* fix: loops

* feat: remove .async. usage

+++ src/database/mongo/sorted/remove.js
@@ -1,45 +1,41 @@
 'use strict';
 
 module.exports = function (db, module) {
-	var helpers = module.helpers.mongo;
+	var helpers = require('../helpers');
 
-	module.sortedSetRemove = function (key, value, callback) {
-		function done(err) {
-			callback(err);
-		}
-		callback = callback || helpers.noop;
+	module.sortedSetRemove = async function (key, value) {
 		if (!key) {
-			return callback();
+			return;
+		}
+		const isValueArray = Array.isArray(value);
+		if (!value || (isValueArray && !value.length)) {
+			return;
 		}
 
-		if (Array.isArray(value)) {
+		if (isValueArray) {
 			value = value.map(helpers.valueToString);
 		} else {
 			value = helpers.valueToString(value);
 		}
 
-		db.collection('objects').deleteMany({
+		await db.collection('objects').deleteMany({
 			_key: Array.isArray(key) ? { $in: key } : key,
-			value: Array.isArray(value) ? { $in: value } : value,
-		}, done);
-	};
-
-	module.sortedSetsRemove = function (keys, value, callback) {
-		callback = callback || helpers.noop;
-		if (!Array.isArray(keys) || !keys.length) {
-			return callback();
-		}
-		value = helpers.valueToString(value);
-
-		db.collection('objects').deleteMany({ _key: { $in: keys }, value: value }, function (err) {
-			callback(err);
+			value: isValueArray ? { $in: value } : value,
 		});
 	};
 
-	module.sortedSetsRemoveRangeByScore = function (keys, min, max, callback) {
-		callback = callback || helpers.noop;
+	module.sortedSetsRemove = async function (keys, value) {
 		if (!Array.isArray(keys) || !keys.length) {
-			return callback();
+			return;
+		}
+		value = helpers.valueToString(value);
+
+		await db.collection('objects').deleteMany({ _key: { $in: keys }, value: value });
+	};
+
+	module.sortedSetsRemoveRangeByScore = async function (keys, min, max) {
+		if (!Array.isArray(keys) || !keys.length) {
+			return;
 		}
 		var query = { _key: { $in: keys } };
 
@@ -51,8 +47,6 @@ module.exports = function (db, module) {
 			query.score.$lte = parseFloat(max);
 		}
 
-		db.collection('objects').deleteMany(query, function (err) {
-			callback(err);
-		});
+		await db.collection('objects').deleteMany(query);
 	};
 };
